added documentation (incomplete) to repo
The documentation added was primarily done for the OCR A-Level Computer Science coursework. The documentation added is an older version as I wasn't able to find the most recent up-to-date document. Documentation is recommended for any A-Level students that wish to see some example work.

Given that its the older version of the document, there may be a few spelling mistakes and/or grammatical errors. It's likely that the documentation will have missing sections and/or spaces. I'll modify the documentation as appropriate if I can find the most recent up-to-date version.

For any A-Level students that are struggling for project layout, I'd recommend looking through this document and taking note of the section structure I implemented.